Added Naming A Pool Feature
Added new feature that allows you to name a pool however you want without any limitations whatsoever. Also, naming the pool using string variable is not going to hurt the performance since the pool still uses enum to work as before. String variable only be effecting the naming of the enum since the prefabs of the pool can be various and have many names to differentiate each other.